requirements.txt
pandas==2.2.2
numpy==1.26.4
mysql-connector-python==8.3.0
python-dotenv==1.0.1
